requirements.txt
black
flake8
isort
loguru~=0.7.2
mkdocs
pip~=24.0
python-dotenv~=1.0.1
tqdm~=4.66.4
typer~=0.12.3
-e .
pandas~=2.2.2
numpy~=1.26.4
matplotlib~=3.9.0
gspread~=6.1.2
pillow~=10.3.0
Jinja2~=3.1.4
typing_extensions~=4.12.2
packaging~=24.1
platformdirs~=4.2.2
requests~=2.32.3